send position in song to companion app, send pitches for beats to main game

+++ UltraStar Play/Packages/playshared/Runtime/Network/Dto/PositionInSongDto.cs
@@ -1,7 +1,7 @@
 public class PositionInSongDto : CompanionAppMessageDto
 {
-    public double SongBpm { get; set; }
-    public double SongGap { get; set; }
+    public float SongBpm { get; set; }
+    public float SongGap { get; set; }
     public int PositionInSongInMillis { get; set; }
 
     public PositionInSongDto() : base(CompanionAppMessageType.PositionInSong) { }
